requirements.txt
selenium
python-telegram-bot
python-dotenv
asyncio